+++ src/main/java/com/kozak/triangles/repositories/LicensesConsignmentRepository.java
@@ -16,25 +16,25 @@ import com.kozak.triangles.entities.LicensesConsignment;
 @Transactional
 public class LicensesConsignmentRepository {
 
-	@PersistenceContext
-	public EntityManager em;
+    @PersistenceContext
+    public EntityManager em;
 
-	public LicensesConsignment addLicenseConsignment(LicensesConsignment licensesConsignment) {
-		return em.merge(licensesConsignment);
-	}
+    public LicensesConsignment addLicenseConsignment(LicensesConsignment licensesConsignment) {
+        return em.merge(licensesConsignment);
+    }
 
-	public void removeConsignment(LicensesConsignment licensesConsignment) {
-		em.remove(licensesConsignment);
-	}
+    public void removeConsignment(LicensesConsignment licensesConsignment) {
+        em.remove(licensesConsignment);
+    }
 
-	/**
-	 * @return список партий лицензий, где дата продажи меньше текущей
-	 */
-	@SuppressWarnings("unchecked")
-	public List<LicensesConsignment> getSoldConsignments(int userId) {
-		String hql = "FROM LicensesConsignment WHERE sellDate < :now AND licenseMarket.userId = :userId";
-		Query query = em.createQuery(hql).setParameter("now", new Date()).setParameter("userId", userId);
-		return query.getResultList();
-	}
+    /**
+     * @return список партий лицензий, где дата продажи меньше текущей
+     */
+    @SuppressWarnings("unchecked")
+    public List<LicensesConsignment> getSoldConsignments(long userId) {
+        String hql = "FROM LicensesConsignment WHERE sellDate < :now AND licenseMarket.userId = :userId";
+        Query query = em.createQuery(hql).setParameter("now", new Date()).setParameter("userId", userId);
+        return query.getResultList();
+    }
 
 }
